% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthday.R
\name{birthday}
\alias{birthday}
\title{Title}
\usage{
birthday(x)
}
\arguments{
\item{x}{the size of the class in students}
}
\value{
the probability of two people having the same birthday in a classs of n students
}
\description{
Title
}
\examples{
birthday(x = 20:25)
}
