% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firedam.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 2 rows and 15 columns:
\describe{
  \item{Distance}{The distance the fire spread}
  \item{Damage}{The damage of the fire}
}
}
\source{
data from Lab 7.
}
\usage{
fire
}
\description{
A dataset containing information about fire damage (from FIREDAM.csv).
}
\keyword{datasets}
