% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstfunction.R
\name{myfirst}
\alias{myfirst}
\title{Myfirst}
\usage{
myfirst(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with components 'x' and 'y', where y is the square of x
}
\description{
Myfirst
}
\examples{
myfirst(1:10)
}
