% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samplingfunction.R
\name{mysample}
\alias{mysample}
\title{Sampling function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Integer. Number of draws in each iteration.}

\item{iter}{Integer. Number of iterations to perform.}

\item{time}{Numeric. Pause (in seconds) between plots.}
}
\value{
A sequence of barplots; invisibly returns the table of counts.
}
\description{
Repeatedly samples numbers 1–10 and shows barplots of frequencies.
}
\examples{
mysample(n=1000, iter=30,time=1)

}
