% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstfunction.R
\name{tss}
\alias{tss}
\title{Calculate Total Sum of Squares}
\usage{
tss(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A value representing the Total Sum of Squares
}
\description{
This function calculates the Total Sum of Squares (TSS) for a given numeric vector
}
\examples{
data_vector <- c(10, 15, 20, 25, 30)
# Calculate the TSS for the vector
tss(data_vector)
}
