% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firedam.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 2 rows and 15 columns:
\describe{
  \item{Variable1}{Description of the first variable}
  \item{Variable2}{Description of the second variable}
  \item{Variable3}{... etc.}
}
}
\source{
Local dataset from ILAS2019 project.
}
\usage{
fire
}
\description{
A dataset containing information about fire damage (from FIREDAM.csv).
}
\keyword{datasets}
