% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overbooking.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on plane}

\item{gamma}{How much "pain" airline is willing to take/tolerated amount of overbooking}

\item{p}{Probability that a person who bought a ticket shows up}
}
\value{
Helps find the optimal amount of tickets to sell on a plane given the number of seats, the probability of a ticket holder not showing up and the airlines tolerance for pain. Creates two plots, one for the discrete distribution and one for the continuous distribution.
}
\description{
Title
}
\examples{
# ntickets(N = 200, gamma = .02, p =.95)
}
